tarea2/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
bootstrap4==0.1.0
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-bootstrap==0.2.4
django-bootstrap4==2.3.1
django-heroku==0.3.1
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
psycopg2==2.8.6
pybase64==1.1.4
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
virtualenv==20.4.3
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
